+++ jdaviz/configs/cubeviz/plugins/viewers.py
@@ -6,6 +6,14 @@ from jdaviz.configs.cubeviz.plugins.mixins import WithSliceIndicator, WithSliceS
 from jdaviz.configs.default.plugins.viewers import JdavizViewerMixin
 from jdaviz.configs.specviz.plugins.viewers import SpecvizProfileView
 from jdaviz.core.freezable_state import FreezableBqplotImageViewerState
+from jdaviz.configs.cubeviz.plugins.cube_listener import CubeListenerData, MINVOL
+import numpy as np
+from astropy import units as u
+
+try:
+    import sounddevice as sd
+except ImportError:
+    pass
 
 __all__ = ['CubevizImageView', 'CubevizProfileView']
 
@@ -30,6 +38,7 @@ class CubevizImageView(JdavizViewerMixin, WithSliceSelection, BqplotImageView):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+
         # provide reference from state back to viewer to use for zoom syncing
         self.state._set_viewer(self)
 
@@ -38,6 +47,13 @@ class CubevizImageView(JdavizViewerMixin, WithSliceSelection, BqplotImageView):
 
         # Hide axes by default
         self.state.show_axes = False
+
+        self.sonified_cube = None
+        self.stream = None
+        self.sonification_wl_bounds = None
+        self.sonification_wl_unit = None
+        self.volume_level = None
+        self.stream_active = True
 
         self.data_menu._obj.dataset.add_filter('is_cube_or_image')
 
@@ -81,6 +97,85 @@ class CubevizImageView(JdavizViewerMixin, WithSliceSelection, BqplotImageView):
                 for layer_state in self.state.layers
                 if hasattr(layer_state, 'layer') and
                 isinstance(layer_state.layer, BaseData)]
+
+    def start_stream(self):
+        if self.stream and not self.stream.closed and self.stream_active:
+            self.stream.start()
+
+    def stop_stream(self):
+        if self.stream and not self.stream.closed and self.stream_active:
+            self.stream.stop()
+
+    def update_sonified_cube(self, x, y):
+        if (not self.sonified_cube or not hasattr(self.sonified_cube, 'newsig') or
+                not hasattr(self.sonified_cube, 'sigcube')):
+            return
+        self.sonified_cube.newsig = self.sonified_cube.sigcube[x, y, :]
+        self.sonified_cube.cbuff = True
+
+    def update_listener_wls(self, w1, w2, wunit):
+        self.sonification_wl_bounds = (w1, w2)
+        self.sonification_wl_unit = wunit
+
+    def update_sound_device(self, device_index):
+        if not self.sonified_cube:
+            return
+
+        self.stop_stream()
+        self.stream = sd.OutputStream(samplerate=self.sample_rate, blocksize=self.buffer_size,
+                                      device=device_index, channels=1, dtype='int16',
+                                      latency='low', callback=self.sonified_cube.player_callback)
+
+    def update_volume_level(self, level):
+        if not self.sonified_cube:
+            return
+        self.volume_level = level
+        self.sonified_cube.atten_level = int(1/np.clip((level/100.)**2, MINVOL, 1))
+
+    def get_sonified_cube(self, sample_rate, buffer_size, device, assidx, ssvidx,
+                          pccut, audfrqmin, audfrqmax, eln):
+        spectrum = self.active_image_layer.layer.get_object(statistic=None)
+        wlens = spectrum.wavelength.to('m').value
+        flux = spectrum.flux.value
+        self.sample_rate = sample_rate
+        self.buffer_size = buffer_size
+
+        if self.sonification_wl_bounds:
+            wl_unit = getattr(u, self.sonification_wl_unit)
+            si_wl_bounds = (self.sonification_wl_bounds * wl_unit).to('m')
+            wdx = np.logical_and(wlens >= si_wl_bounds[0].value,
+                                 wlens <= si_wl_bounds[1].value)
+            wlens = wlens[wdx]
+            flux = flux[:, :, wdx]
+
+        pc_cube = np.percentile(np.nan_to_num(flux), np.clip(pccut, 0, 99), axis=-1)
+
+        # clip zeros and remove NaNs
+        clipped_arr = np.nan_to_num(np.clip(flux, 0, np.inf), copy=False)
+
+        # make a rough white-light image from the clipped array
+        whitelight = np.expand_dims(clipped_arr.sum(-1), axis=2)
+
+        # subtract any percentile cut
+        clipped_arr -= np.expand_dims(pc_cube, axis=2)
+
+        # and re-clip
+        clipped_arr = np.clip(clipped_arr, 0, np.inf)
+
+        self.sonified_cube = CubeListenerData(clipped_arr ** assidx, wlens, duration=0.8,
+                                              samplerate=sample_rate, buffsize=buffer_size,
+                                              wl_bounds=self.sonification_wl_bounds,
+                                              wl_unit=self.sonification_wl_unit,
+                                              audfrqmin=audfrqmin, audfrqmax=audfrqmax,
+                                              eln=eln, vol=self.volume_level)
+        self.sonified_cube.sonify_cube()
+        self.sonified_cube.sigcube = (
+                self.sonified_cube.sigcube * pow(whitelight / whitelight.max(),
+                                                 ssvidx)).astype('int16')
+        self.stream = sd.OutputStream(samplerate=sample_rate, blocksize=buffer_size, device=device,
+                                      channels=1, dtype='int16', latency='low',
+                                      callback=self.sonified_cube.player_callback)
+        self.sonified_cube.cbuff = True
 
 
 @viewer_registry("cubeviz-profile-viewer", label="Profile 1D (Cubeviz)")